requirements.txt
torch
matplotlib
pillow
numpy